chore: relocate init script - update paths
