+++ contract/scripts/init-orca.js
@@ -21,8 +21,8 @@ import { parseArgs } from 'node:util';
 import {
   getManifestForOrca,
   startOrcaContract,
-} from '../../src/orca.proposal.js';
-import { makeAgd } from '../../tools/agd-lib.js';
+} from '../src/orca.proposal.js';
+import { makeAgd } from '../tools/agd-lib.js';
 
 /**
  * @import {CoreEvalBuilder} from '@agoric/deploy-script-support/src/externalTypes.js'
@@ -43,12 +43,12 @@ export const defaultProposalBuilder = async (
   { chainDetails },
 ) => {
   return harden({
-    sourceSpec: '../../src/orca.proposal.js',
+    sourceSpec: '../src/orca.proposal.js',
     getManifestCall: [
       getManifestForOrca.name,
       {
         installKeys: {
-          orca: publishRef(install('../../src/orca.contract.js')),
+          orca: publishRef(install('../src/orca.contract.js')),
         },
         chainDetails,
       },
